+++ gamemodes/terrortown/entities/entities/ttt_random_ammo.lua
@@ -5,3 +5,13 @@
 
 ENT.Type = "point"
 ENT.Base = "base_point"
+
+---
+-- @note Only used to forceSpawn ammo after map cleanup 
+-- otherwise these entities are only used to mark the spots for random ammo spawns
+-- @realm shared
+function ENT:Initialize()
+	if entspawn.IsForcedRandomSpawnEnabled() then
+		entspawn.SpawnRandomAmmo(self)
+	end
+end