+++ gamemodes/terrortown/gamemode/client/cl_targetid.lua
@@ -118,7 +118,7 @@ function DrawOverheadRoleIcon(ply, ricon, rcolor)
 
 	-- draw icon
 	render.SetMaterial(ricon)
-	render.DrawQuadEasy(Vector(pos.x, pos.y, pos.z + 0.5), dir, 8, 8, COLOR_WHITE, 180)
+	render.DrawQuadEasy(Vector(pos.x, pos.y, pos.z + 0.5), dir, 8, 8, util.GetDefaultColor(rcolor), 180)
 
 	-- stop linear filter
 	render.PopFilterMag()
@@ -172,13 +172,20 @@ function GM:PostDrawTranslucentRenderables(bDrawingDepth, bDrawingSkybox)
 		local ply = plys[i]
 		local rd = ply:GetSubRoleData()
 
-		if ply:IsActive()
+		local shouldDrawDefault = ply:IsActive()
 			and ply:HasRole()
 			and (not client:IsActive() or ply:IsInTeam(client) or rd.isPublicRole)
 			and not rd.avoidTeamIcons
-		then
-			DrawOverheadRoleIcon(ply, rd.iconMaterial, ply:GetRoleColor())
-		end
+
+		---
+		-- @realm client
+		local shouldDraw, material, color = hook.Run("TTT2ModifyOverheadIcon", ply, shouldDrawDefault)
+
+		if shouldDraw == false
+			or not shouldDrawDefault and not (material and color)
+		then continue end
+
+		DrawOverheadRoleIcon(ply, material or rd.iconMaterial, color or ply:GetRoleColor())
 	end
 end
 
@@ -434,5 +441,18 @@ end
 -- @hook
 -- @realm client
 function GM:TTTModifyTargetTracedata(startpos, endpos)
+
+end
+
+---
+-- Can be used to modify, disable or add an overhead icon of a player.
+-- @param Player ply The player whose overhead icon should be modified
+-- @param boolean shouldDrawDefault Defines if the overhead icon would draw normally
+-- @return boolean Return false to prevent overhead icon draw
+-- @return Material The material for the overhead icon (normally the role icon)
+-- @return Color The color of the overhead icon backdrop
+-- @hook
+-- @realm client
+function GM:TTT2ModifyOverheadIcon(ply, shouldDrawDefault)
 
 end