+++ gamemodes/terrortown/gamemode/server/sv_main.lua
@@ -336,7 +336,6 @@ function GM:Initialize()
 
 	self.DamageLog = {}
 	self.LastRole = {}
-	self.playercolor = COLOR_WHITE
 
 	-- Delay reading of cvars until config has definitely loaded
 	self.cvar_init = false
@@ -426,34 +425,6 @@ function GM:InitPostEntity()
 	-- load all HUD elements
 	hudelements.OnLoaded()
 
-	local itms = items.GetList()
-
-	local isSqlTableCreated = sql.CreateSqlTable("ttt2_items", ShopEditor.savingKeys)
-
-	for i = 1, #itms do
-		local eq = itms[i]
-
-		InitDefaultEquipment(eq)
-		ShopEditor.InitDefaultData(eq)
-
-		if isSqlTableCreated then
-			local name = GetEquipmentFileName(WEPS.GetClass(eq))
-			local loaded, changed = sql.Load("ttt2_items", name, eq, ShopEditor.savingKeys)
-
-			if not loaded then
-				sql.Init("ttt2_items", name, eq, ShopEditor.savingKeys)
-			elseif changed then
-				CHANGED_EQUIPMENT[#CHANGED_EQUIPMENT + 1] = {name, eq}
-			end
-		end
-
-		CreateEquipment(eq) -- init items
-
-		eq.CanBuy = {} -- reset normal items equipment
-
-		eq:Initialize()
-	end
-
 	local sweps = weapons.GetList()
 
 	for i = 1, #sweps do
@@ -520,6 +491,10 @@ function GM:InitPostEntity()
 	-- initialize playermodel database
 	playermodels.Initialize()
 	self.playermodel = playermodels.GetRandomPlayerModel()
+
+	-- set the default random playermodel
+	self.playermodel = playermodels.GetRandomPlayerModel()
+	self.playercolor = COLOR_WHITE
 
 	timer.Simple(0, function()
 		addonChecker.Check()
@@ -884,6 +859,9 @@ function GM:PreCleanupMap()
 	ents.TTT.FixParentedPreCleanup()
 
 	entityOutputs.CleanUp()
+
+	-- While cleaning up the map, disable random weapons directly spawning
+	entspawn.SetForcedRandomSpawn(false)
 end
 
 ---
@@ -898,6 +876,11 @@ function GM:PostCleanupMap()
 	entityOutputs.SetUp()
 
 	entspawn.HandleSpawns()
+
+	-- After map cleanup enable 'env_entity_maker'-entities to force spawn random weapons and ammo
+	-- This is necessary for maps like 'ttt_lttp_kakariko_a5', that only initialize 'ttt_random_weapon'-entities
+	-- after destroying vases and were therefore not affected by our entspawn-system
+	entspawn.SetForcedRandomSpawn(true)
 	---
 	-- @realm server
 	hook.Run("TTT2PostCleanupMap")
@@ -1360,6 +1343,9 @@ function GM:OnReloaded()
 	-- reload entity spawns from file
 	entspawnscript.OnLoaded()
 
+	-- load all items
+	items.OnLoaded()
+
 	-- load all HUDs
 	huds.OnLoaded()
 
@@ -1368,6 +1354,10 @@ function GM:OnReloaded()
 
 	-- reload everything from the playermodels
 	playermodels.Initialize()
+
+	-- set the default random playermodel
+	self.playermodel = playermodels.GetRandomPlayerModel()
+	self.playercolor = COLOR_WHITE
 
 	---
 	-- @realm shared
@@ -1486,6 +1476,8 @@ end
 ---
 -- Can be used to modify the table of teams with alive players. This hook is
 -- used in the default win condition.
+-- @note A dead player that is revived is counted as alive as well if the revival mode
+-- ist set to blocking mode.
 -- @param table alives The table of teams which have at least one player still alive
 -- @hook
 -- @realm server
@@ -1534,34 +1526,39 @@ function GM:TTTCheckForWin()
 	end
 
 	if self.MapWin ~= WIN_NONE then -- a role wins
-		local mw = self.MapWin
+		local mapWin = self.MapWin
 
 		self.MapWin = WIN_NONE
 
-		return mw
+		return mapWin
 	end
 
-	local alive = {}
+	local aliveTeams = {}
 	local plys = player.GetAll()
 
 	for i = 1, #plys do
-		local v = plys[i]
-		local tm = v:GetTeam()
+		local ply = plys[i]
+		local team = ply:GetTeam()
 
-		if (v:IsTerror() or v:IsBlockingRevival()) and not v:GetSubRoleData().preventWin and tm ~= TEAM_NONE then
-			alive[#alive + 1] = tm
+		if (ply:IsTerror() or ply:IsBlockingRevival()) and not ply:GetSubRoleData().preventWin and team ~= TEAM_NONE then
+			aliveTeams[#aliveTeams + 1] = team
+		end
+
+		-- special case: The revival blocks the round end
+		if ply:GetRevivalBlockMode() == REVIVAL_BLOCK_ALL then
+			return WIN_NONE
 		end
 	end
 
 	---
 	-- @realm server
-	hook.Run("TTT2ModifyWinningAlives", alive)
+	hook.Run("TTT2ModifyWinningAlives", aliveTeams)
 
 	local checkedTeams = {}
 	local b = 0
 
-	for i = 1, #alive do
-		local team = alive[i]
+	for i = 1, #aliveTeams do
+		local team = aliveTeams[i]
 
 		if team == TEAM_NONE then continue end
 
@@ -1580,7 +1577,7 @@ function GM:TTTCheckForWin()
 	if b > 1 then -- if >= 2 teams alive: no one wins
 		return WIN_NONE -- early out
 	elseif b == 1 then -- just 1 team is alive
-		return alive[1]
+		return aliveTeams[1]
 	else -- rare case: nobody is alive, e.g. because of an explosion
 		return TEAM_NONE -- none_win
 	end