+++ lua/autorun/server/ttt2_force_download.lua
@@ -111,6 +111,8 @@ resource.AddFile("materials/vgui/ttt/perks/hud_noenergydmg.png")
 resource.AddFile("materials/vgui/ttt/icon_nohazarddmg.vmt")
 resource.AddFile("materials/vgui/ttt/perks/hud_nohazarddmg.png")
 
+resource.AddFile("materials/vgui/ttt/icon_magneto_stick.vmt")
+
 -- pickup symbols
 resource.AddFile("materials/vgui/ttt/pickup/icon_heavy.png")
 resource.AddFile("materials/vgui/ttt/pickup/icon_pistol.png")