+++ lua/terrortown/entities/roles/ttt_role_base/shared.lua
@@ -40,9 +40,29 @@ ROLE.score = {
 	-- round ended with nobody winning, usually a negative number.
 	timelimitMultiplier = 0,
 
-	-- the amount of points gained by killing yourself. Should be a
+	-- The amount of points gained by killing yourself. Should be a
 	-- negative number for most roles.
 	suicideMultiplier = -1
+}
+
+ROLE.karma = {
+	-- The multiplier that is used to calculate the Karma penalty for a team kill.
+	-- Keep in mind that the game will increase the multiplier further if it was avoidable
+	-- like a kill on a public policing role.
+	teamKillPenaltyMultiplier = 1,
+
+	-- The multiplier that is used to calculate the Karma penalty for team damage.
+	-- Keep in mind that the game will increase the multiplier further if it was avoidable
+	-- like damage applied to a public policing role.
+	teamHurtPenaltyMultiplier = 1,
+
+	-- The multiplier that is used to change the Karma given to the killer if a player
+	-- from an enemy team is killed.
+	enemyKillBonusMultiplier = 1,
+
+	-- The multiplier that is used to change the Karma given to the attacker if a player
+	-- from an enemy team is damaged.
+	enemyHurtBonusMultiplier = 1,
 }
 
 ROLE.conVarData = {
@@ -59,7 +79,7 @@ ROLE.conVarData = {
 	minKarma = 0,
 
 	-- Defines if the role has access to traitor buttons.
-	traitorButton = 1,
+	traitorButton = 0,
 
 	-- Sets the amount of credits the role is starting with.
 	credits = 0,
@@ -84,15 +104,19 @@ ROLE.conVarData = {
 -- selection process.
 ROLE.notSelectable = false
 
--- This variable can be used to add roles that can see the role of the
+-- This variable can be used to add teams that can see the role of the
 -- player with the role defined in this file. While this table can be updated
 -- on runtime, it is strongly advised against. Use custom hook based
 -- syncing for specific syncing.
 ROLE.visibleForTeam = {}
 
+-- This variable can be used to add roles that can can be seen by the role
+-- defined in this file.
+ROLE.networkRoles = {}
+
 -- If set to true, this role doesn't know about their teammates. A normal innocent
 -- for example knows that they are in team innocent, but doesn't know who else is.
--- If set to false, this player knows about the role of all their teamm ates.
+-- If set to false, this player knows about the role of all their team mates.
 ROLE.unknownTeam = false
 
 -- This can be used to force prevent picking up credits from corpses. Keep in mind
@@ -120,6 +144,29 @@ ROLE.isPublicRole = false
 -- A policing role is a role that works like a detective. They can be called to
 -- a corpse and evil roles can be rewarded more points for them being killed.
 ROLE.isPolicingRole = false
+
+-- Omniscient roles are able to see missing in action players and therefore the haste
+-- mode timer as well. This is mostly traitor-like behaviour.
+ROLE.isOmniscientRole = false
+
+-- If this is set to true, the role is unable to send messages in the team chat. If this is
+-- set to false, it still could mean that the player is unable to use the team chat. If the
+-- role flag `.unknownTeam` is set, the team chat can't be used either.
+ROLE.disabledTeamChat = false
+
+-- If this is set to true, the geiven role is unable to receive team chat messages.
+ROLE.disabledTeamChatRecv = false
+
+-- By setting this to true, the role is unable to write in the general chat.
+ROLE.disabledGeneralChat = false
+
+-- If this is set to true, the role is unable to speak in the team voice chat. If this is
+-- set to false, it still could mean that the player is unable to use the team chat. If the
+-- role flag `.unknownTeam` is set, the team chat can't be used either.
+ROLE.disabledTeamVoice = false
+
+-- If this is set to true, the given role is unable to hear team voice chat.
+ROLE.disabledTeamVoiceRecv = false
 
 ---
 -- This function is called before initializing a @{ROLE}, but after all