+++ lua/ttt2/libraries/entspawn.lua
@@ -17,7 +17,65 @@ local mathMin = math.min
 local timerCreate = timer.Create
 local timerRemove = timer.Remove
 
+local allowForcedRandomSpawn = false
+
 entspawn = entspawn or {}
+
+---
+-- Enable or disable forced random spawns for 'env_entity_maker' spawning non available
+-- random spawns at map start.
+-- @note see: https://developer.valvesoftware.com/wiki/Env_entity_maker
+-- @param bool enable The state to set it to.
+-- @realm server
+function entspawn.SetForcedRandomSpawn(enable)
+	allowForcedRandomSpawn = enable
+end
+
+---
+-- To check if forced random spawns are available.
+-- @return bool if forced random spawns are enabled
+-- @realm server
+function entspawn.IsForcedRandomSpawnEnabled()
+	return allowForcedRandomSpawn
+end
+
+---
+-- Spawns a random weapon with the given data of the random spawn entity.
+-- @param Entity ent the entity holding the random weapon data
+-- @realm server
+function entspawn.SpawnRandomWeapon(ent)
+	local spawns = {
+		[SPAWN_TYPE_WEAPON] = {
+			[1] = {
+				pos = ent:GetPos(),
+				ang = ent:GetAngles(),
+				ammo = ent.autoAmmoAmount or 0
+				}
+			}
+		}
+
+	local wepsForTypes, weps = WEPS.GetWeaponsForSpawnTypes()
+	entspawn.SpawnEntities(spawns, wepsForTypes, weps, WEAPON_TYPE_RANDOM)
+end
+
+---
+-- Spawns a random ammo box with the given data of the random spawn entity.
+-- @param Entity ent the entity holding the random ammo data
+-- @realm server
+function entspawn.SpawnRandomAmmo(ent)
+	local spawns = {
+		[SPAWN_TYPE_AMMO] = {
+			[1] = {
+				pos = ent:GetPos(),
+				ang = ent:GetAngles(),
+				ammo = 1
+				}
+			}
+		}
+
+	local ammoForTypes, ammo = WEPS.GetAmmoForSpawnTypes()
+	entspawn.SpawnEntities(spawns, ammoForTypes, ammo, AMMO_TYPE_RANDOM)
+end
 
 local function RemoveEntities(entTable, spawnTable, spawnType)
 	local useDefaultSpawns = not entspawnscript.ShouldUseCustomSpawns()